feature/crud add trip model add search + filtering

+++ src/main/java/ru/taxi/adminpanel/backend/dto/RecordDTO.java
@@ -5,7 +5,6 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-import javax.persistence.Column;
 import java.time.LocalDateTime;
 
 @Data
@@ -22,5 +21,5 @@ public class RecordDTO {
 
     private LocalDateTime end;
 
-    private Long price;
+    private Double price;
 }
